backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.23
psycopg2-binary==2.9.7
alembic==1.12.1
pydantic==2.5.0
pydantic-settings==2.1.0
celery==5.3.4
flower==2.0.1
redis==4.5.5
pymupdf==1.23.8
openai==1.3.0
prometheus-client==0.20.0
python-multipart==0.0.6
python-dotenv==1.0.0
tqdm==4.66.1
aiofiles==23.2.1
python-bidi==0.4.2
arabic-reshaper==3.0.0
tiktoken==0.5.1
langdetect==1.0.9
pdfplumber==0.10.3
reportlab==4.0.7
python-docx==1.1.0

# Authentication and Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
email-validator==2.1.0

# Natural Language Processing
nltk==3.8.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1
pytest-mock==3.12.0
httpx==0.25.2
factory-boy==3.3.0
faker==20.1.0

# Linting and Static Analysis
mypy==1.7.1
ruff==0.1.6
black==23.11.0
isort==5.12.0
flake8==6.1.0
bandit==1.7.5

# Import Analysis
import-linter==1.9.0
unimport==1.0.0
